changing error message format for E0611
