+++ src/librustc/infer/error_reporting/anon_anon_conflict.rs
@@ -12,10 +12,9 @@
 //! where both the regions are anonymous.
 use hir;
 use infer::InferCtxt;
-use ty::{self, Region};
+use ty;
 use infer::region_inference::RegionResolutionError::*;
 use infer::region_inference::RegionResolutionError;
-use hir::map as hir_map;
 use middle::resolve_lifetime as rl;
 use hir::intravisit::{self, Visitor, NestedVisitorMap};
 
@@ -86,9 +85,6 @@ impl<'a, 'gcx, 'tcx> InferCtxt<'a, 'gcx, 'tcx> {
                 return false;
             }
 
-
-
-
             if anon_arg1 == anon_arg2 {
                 (format!(" with one lifetime"), format!(" into the other"))
             } else {
@@ -119,81 +115,6 @@ impl<'a, 'gcx, 'tcx> InferCtxt<'a, 'gcx, 'tcx> {
         return true;
 
     }
-
-    /// This function calls the `visit_ty` method for the parameters
-    /// corresponding to the anonymous regions. The `nested_visitor.found_type`
-    /// contains the anonymous type.
-    ///
-    /// # Arguments
-    ///
-    /// region - the anonymous region corresponding to the anon_anon conflict
-    /// br - the bound region corresponding to the above region which is of type `BrAnon(_)`
-    ///
-    /// # Example
-    /// ```
-    /// fn foo(x: &mut Vec<&u8>, y: &u8)
-    ///    { x.push(y); }
-    /// ```
-    /// The function returns the nested type corresponding to the anonymous region
-    /// for e.g. `&u8` and Vec<`&u8`.
-    pub fn find_anon_type(&self, region: Region<'tcx>, br: &ty::BoundRegion) -> Option<(&hir::Ty)> {
-        if let Some(anon_reg) = self.is_suitable_anonymous_region(region, true) {
-            let (def_id, _) = anon_reg;
-            if let Some(node_id) = self.tcx.hir.as_local_node_id(def_id) {
-                let ret_ty = self.tcx.type_of(def_id);
-                if let ty::TyFnDef(_, _) = ret_ty.sty {
-                    if let hir_map::NodeItem(it) = self.tcx.hir.get(node_id) {
-                        if let hir::ItemFn(ref fndecl, _, _, _, _, _) = it.node {
-                            return fndecl
-                                       .inputs
-                                       .iter()
-                                       .filter_map(|arg| {
-                                                       self.find_visitor_found_type(&**arg, br)
-                                                   })
-                                       .next();
-                        }
-                    } else if let hir_map::NodeTraitItem(it) = self.tcx.hir.get(node_id) {
-                        if let hir::TraitItemKind::Method(ref fndecl, _) = it.node {
-                            return fndecl
-                                       .decl
-                                       .inputs
-                                       .iter()
-                                       .filter_map(|arg| {
-                                                       self.find_visitor_found_type(&**arg, br)
-                                                   })
-                                       .next();
-                        }
-                    } else if let hir_map::NodeImplItem(it) = self.tcx.hir.get(node_id) {
-                        if let hir::ImplItemKind::Method(ref fndecl, _) = it.node {
-                            return fndecl
-                                       .decl
-                                       .inputs
-                                       .iter()
-                                       .filter_map(|arg| {
-                                                       self.find_visitor_found_type(&**arg, br)
-                                                   })
-                                       .next();
-                        }
-                    }
-                }
-            }
-        }
-        None
-    }
-
-    fn find_visitor_found_type(&self,
-                               arg: &'gcx hir::Ty,
-                               br: &ty::BoundRegion)
-                               -> Option<(&'gcx hir::Ty)> {
-        let mut nested_visitor = FindNestedTypeVisitor {
-            infcx: &self,
-            hir_map: &self.tcx.hir,
-            bound_region: *br,
-            found_type: None,
-        };
-        nested_visitor.visit_ty(arg);
-        nested_visitor.found_type
-    }
 }
 
 // The FindNestedTypeVisitor captures the corresponding `hir::Ty` of the
@@ -203,15 +124,15 @@ impl<'a, 'gcx, 'tcx> InferCtxt<'a, 'gcx, 'tcx> {
 // walk the types like &mut Vec<&u8> and &u8 looking for the HIR
 // where that lifetime appears. This allows us to highlight the
 // specific part of the type in the error message.
-struct FindNestedTypeVisitor<'a, 'gcx: 'a + 'tcx, 'tcx: 'a> {
-    infcx: &'a InferCtxt<'a, 'gcx, 'tcx>,
-    hir_map: &'a hir::map::Map<'gcx>,
+pub struct FindNestedTypeVisitor<'a, 'gcx: 'a + 'tcx, 'tcx: 'a> {
+    pub infcx: &'a InferCtxt<'a, 'gcx, 'tcx>,
+    pub hir_map: &'a hir::map::Map<'gcx>,
     // The bound_region corresponding to the Refree(freeregion)
     // associated with the anonymous region we are looking for.
-    bound_region: ty::BoundRegion,
+    pub bound_region: ty::BoundRegion,
     // The type where the anonymous lifetime appears
     // for e.g. Vec<`&u8`> and <`&u8`>
-    found_type: Option<&'gcx hir::Ty>,
+    pub found_type: Option<&'gcx hir::Ty>,
 }
 
 impl<'a, 'gcx, 'tcx> Visitor<'gcx> for FindNestedTypeVisitor<'a, 'gcx, 'tcx> {
