+++ src/test/ui/lifetime-errors/ex1-return-one-existing-name-if-else-using-impl-2.stderr
@@ -2,7 +2,7 @@ error[E0621]: explicit lifetime required in the type of `x`
   --> $DIR/ex1-return-one-existing-name-if-else-using-impl-2.rs:14:15
    |
 13 | fn foo<'a>(x: &i32, y: &'a i32) -> &'a i32 {
-   |            - consider changing the type of `x` to `&'a i32`
+   |               ---- consider changing the type of `x` to `&'a i32`
 14 |    if x > y { x } else { y }
    |               ^ lifetime `'a` required
