+++ src/test/ui/lifetime-errors/ex1-return-one-existing-name-if-else-using-impl-3.stderr
@@ -2,7 +2,7 @@ error[E0621]: explicit lifetime required in the type of `x`
   --> $DIR/ex1-return-one-existing-name-if-else-using-impl-3.rs:18:36
    |
 16 |   fn foo<'a>(&'a self, x: &i32) -> &i32 {
-   |                        - consider changing the type of `x` to `&'a i32`
+   |                           ---- consider changing the type of `x` to `&'a i32`
 17 | 
 18 |     if true { &self.field } else { x }
    |                                    ^ lifetime `'a` required
