+++ src/test/ui/lifetime-errors/ex2a-push-one-existing-name-2.stderr
@@ -2,7 +2,7 @@ error[E0621]: explicit lifetime required in the type of `x`
   --> $DIR/ex2a-push-one-existing-name-2.rs:16:12
    |
 15 | fn foo<'a>(x: Ref<i32>, y: &mut Vec<Ref<'a, i32>>) {
-   |            - consider changing the type of `x` to `Ref<'a, i32>`
+   |               -------- consider changing the type of `x` to `Ref<'a, i32>`
 16 |     y.push(x);
    |            ^ lifetime `'a` required
