+++ src/test/ui/lifetime-errors/ex2a-push-one-existing-name.stderr
@@ -2,7 +2,7 @@ error[E0621]: explicit lifetime required in the type of `y`
   --> $DIR/ex2a-push-one-existing-name.rs:16:12
    |
 15 | fn foo<'a>(x: &mut Vec<Ref<'a, i32>>, y: Ref<i32>) {
-   |                                       - consider changing the type of `y` to `Ref<'a, i32>`
+   |                                          -------- consider changing the type of `y` to `Ref<'a, i32>`
 16 |     x.push(y);
    |            ^ lifetime `'a` required
